pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amadeus>=8.0.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "openmeteo-requests>=1.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "reportlab>=4.0.0",
    "requests>=2.32.5",
    "sendgrid>=6.12.4",
    "serpapi>=0.1.5",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
